move ParseServiceProvider and twilioClient into providers dir. remove errors/index.js and use ExtendableErrorClass package instead
