--- src/errors/index.js
@@ -1,18 +1,0 @@
-
-
-export class ExtendableError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = this.constructor.name;
-    this.message = message;
-    if (typeof Error.captureStackTrace === 'function') {
-      Error.captureStackTrace(this, this.constructor);
-    } else {
-      this.stack = (new Error(message)).stack;
-    }
-  }
-}
-
-export class ArgumentTypeError extends ExtendableError {}
-export class RequestBodyError extends ExtendableError {}
-export class ObjectNotFoundError extends ExtendableError {}
