+++ src/providers/TwilioServiceProvider.js
@@ -9,11 +9,11 @@ const {
 
 // Throw if required env vars are not declared before trying to build instance
 if (!TWILIO_ACCOUNT_SID || !TWILIO_AUTH_TOKEN) {
-  throw new Error('TWILIO_ACCOUNT_SID and TWILIO_AUTH_TOKEN environment variables are required');
+  throw new Error('[2xYv/ijM] TWILIO_ACCOUNT_SID and TWILIO_AUTH_TOKEN environment variables are required');
 }
 
 
-const twilioClient = new twilio(TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN);
+const Twilio = new twilio(TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN);
 
 
-export default twilioClient;
+export default Twilio;
